server/requirements.txt
torch
torchvision
av
#fastapi
#uvicorn
#cudatoolkit=11.0
opencv-python
matplotlib
numpy
Pillow
dadaptation
openmim
torchserve
torch-model-archiver
torch-workflow-archiver